requirements.txt
python-telegram-bot==20.7
APScheduler==3.10.4
feedparser==6.0.11
aiohttp==3.9.1
